AI-Models/Critique-Models/description_generator/requirements.txt
fastapi
uvicorn
torch==2.0.1+cpu
torchvision==0.15.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html
pillow
numpy<2
python-multipart
transformers==4.31.0  # <-- downgrade to this version or similar
requests
timm